AVX512: scatter update loop
After updating v16 we updated the counter in a for loop (unrolled anyway by a compiler).
This change interleaves updates of the counters with population count instruciton.
The update is done for *the previous value of v16*, thus after the main loop completes,
we have to include the last v16's value.